requirements.txt
# Core
streamlit==1.28.0
numpy==1.24.3
python-dotenv==1.0.0

# Embeddings
sentence-transformers==2.2.2

# Vector Database
faiss-cpu==1.7.4

# OCR
easyocr==1.7.0
torch==2.0.1
pdf2image==1.16.3
opencv-python==4.8.1.78
pytesseract==0.3.10

# Document Processing - FIXED VERSIONS
PyPDF2==3.0.1
python-docx==0.8.11  # This version works
openpyxl==3.1.2
pandas==2.0.3
python-pptx==0.6.21

# Gemini
google-genai==0.2.0

# Image Processing
Pillow==10.0.0

# Utilities
python-magic-bin==0.4.14